+++ assets/js/dashboard/query.js
@@ -1,18 +1,21 @@
 import React from 'react'
 import { Link, withRouter } from 'react-router-dom'
+import JsonURL from '@jsonurl/jsonurl'
+import { PlausibleSearchParams, updatedQuery } from './util/url'
 import { nowForSite } from './util/date'
 import * as storage from './util/storage'
 import { COMPARISON_DISABLED_PERIODS, getStoredComparisonMode, isComparisonEnabled, getStoredMatchDayOfWeek } from './comparison-input'
 
 import dayjs from 'dayjs';
 import utc from 'dayjs/plugin/utc';
+import { parseLegacyFilter, parseLegacyPropsFilter } from './util/filters'
 
 dayjs.extend(utc)
 
 const PERIODS = ['realtime', 'day', 'month', '7d', '30d', '6mo', '12mo', 'year', 'all', 'custom']
 
 export function parseQuery(querystring, site) {
-  const q = new URLSearchParams(querystring)
+  const q = new PlausibleSearchParams(querystring)
   let period = q.get('period')
   const periodKey = `period__${site.domain}`
 
@@ -40,66 +43,88 @@ export function parseQuery(querystring, site) {
     match_day_of_week: matchDayOfWeek == 'true',
     with_imported: q.get('with_imported') ? q.get('with_imported') === 'true' : true,
     experimental_session_count: q.get('experimental_session_count'),
-    filters: {
-      'goal': q.get('goal'),
-      'props': JSON.parse(q.get('props')),
-      'source': q.get('source'),
-      'utm_medium': q.get('utm_medium'),
-      'utm_source': q.get('utm_source'),
-      'utm_campaign': q.get('utm_campaign'),
-      'utm_content': q.get('utm_content'),
-      'utm_term': q.get('utm_term'),
-      'referrer': q.get('referrer'),
-      'screen': q.get('screen'),
-      'browser': q.get('browser'),
-      'browser_version': q.get('browser_version'),
-      'os': q.get('os'),
-      'os_version': q.get('os_version'),
-      'country': q.get('country'),
-      'region': q.get('region'),
-      'city': q.get('city'),
-      'page': q.get('page'),
-      'hostname': q.get('hostname'),
-      'entry_page': q.get('entry_page'),
-      'exit_page': q.get('exit_page')
-    }
+    filters: JsonURL.parse(q.get('filters')) || [],
+    labels: JsonURL.parse(q.get('labels')) || {}
   }
-}
-
-export function appliedFilters(query) {
-  const propKeys = Object.entries(query.filters.props || {})
-    .map(([key, value]) => ({ key, value, filterType: 'props' }))
-
-  return Object.entries(query.filters)
-    .map(([key, value]) => ({ key, value, filterType: key }))
-    .filter(({ key, value }) => key !== 'props' && !!value)
-    .concat(propKeys)
-}
-
-function generateQueryString(data) {
-  const query = new URLSearchParams(window.location.search)
-  Object.keys(data).forEach(key => {
-    if (!data[key]) {
-      query.delete(key)
-      return
-    }
-
-    query.set(key, data[key])
-  })
-  return query.toString()
 }
 
 export function navigateToQuery(history, queryFrom, newData) {
   // if we update any data that we store in localstorage, make sure going back in history will
   // revert them
   if (newData.period && newData.period !== queryFrom.period) {
-    const replaceQuery = new URLSearchParams(window.location.search)
-    replaceQuery.set('period', queryFrom.period)
-    history.replace({ search: replaceQuery.toString() })
+    history.replace({ search: updatedQuery({ period: queryFrom.period}) })
   }
 
   // then push the new query to the history
-  history.push({ search: generateQueryString(newData) })
+  history.push({ search: updatedQuery(newData) })
+}
+
+const LEGACY_URL_PARAMETERS = {
+  'goal': null,
+  'source': null,
+  'utm_medium': null,
+  'utm_source': null,
+  'utm_campaign': null,
+  'utm_content': null,
+  'utm_term': null,
+  'referrer': null,
+  'screen': null,
+  'browser': null,
+  'browser_version': null,
+  'os': null,
+  'os_version': null,
+  'country': 'country_labels',
+  'region': 'region_labels',
+  'city': 'city_labels',
+  'page': null,
+  'hostname': null,
+  'entry_page': null,
+  'exit_page': null,
+}
+
+// Called once when dashboard is loaded load. Checks whether old filter style is used and if so,
+// updates the filters and updates location
+export function filtersBackwardsCompatibilityRedirect() {
+  const q = new PlausibleSearchParams(window.location.search)
+  const entries = Array.from(q.entries())
+
+  // New filters are used - no need to do anything
+  if (q.get("filters")) {
+    return
+  }
+
+  let filters = []
+  let labels = {}
+
+  for (const [key, value] of entries) {
+    if (LEGACY_URL_PARAMETERS.hasOwnProperty(key)) {
+      const filter = parseLegacyFilter(key, value)
+      filters.push(filter)
+      q.delete(key)
+
+      const labelsKey = LEGACY_URL_PARAMETERS[key]
+      if (labelsKey && q.get(labelsKey)) {
+        const clauses = filter[2]
+        const labelsValues = q.get(labelsKey).split('|').filter(label => !!label)
+        const newLabels = Object.fromEntries(clauses.map((clause, index) => [clause, labelsValues[index]]))
+
+        labels = Object.assign(labels, newLabels)
+        q.delete(labelsKey)
+      }
+    }
+  }
+
+  if (q.get('props')) {
+    filters.push(...parseLegacyPropsFilter(q.get('props')))
+    q.delete('props')
+  }
+
+  if (filters.length > 0) {
+    q.set('filters', filters)
+    q.set('labels', labels)
+
+    history.pushState({}, null, `${window.location.pathname}?${q.toString()}`)
+  }
 }
 
 class QueryLink extends React.Component {
@@ -119,7 +144,7 @@ class QueryLink extends React.Component {
     return (
       <Link
         {...props}
-        to={{ pathname: window.location.pathname, search: generateQueryString(to) }}
+        to={{ pathname: window.location.pathname, search: updatedQuery(to) }}
         onClick={this.onClick}
       />
     )
@@ -136,7 +161,7 @@ function QueryButton({ history, query, to, disabled, className, children, onClic
         event.preventDefault()
         navigateToQuery(history, query, to)
         if (onClick) onClick(event)
-        history.push({ pathname: window.location.pathname, search: generateQueryString(to) })
+        history.push({ pathname: window.location.pathname, search: updatedQuery(to) })
       }}
       type="button"
       disabled={disabled}
@@ -148,13 +173,3 @@ function QueryButton({ history, query, to, disabled, className, children, onClic
 
 const QueryButtonWithRouter = withRouter(QueryButton)
 export { QueryButtonWithRouter as QueryButton };
-
-export function eventName(query) {
-  if (query.filters.goal) {
-    if (query.filters.goal.startsWith('Visit ')) {
-      return 'pageviews'
-    }
-    return 'events'
-  }
-  return 'pageviews'
-}