+++ test/plausible_web/controllers/site/membership_controller_test.exs
@@ -1,9 +1,12 @@
 defmodule PlausibleWeb.Site.MembershipControllerTest do
+  use Plausible
   use PlausibleWeb.ConnCase
   use Plausible.Repo
   use Bamboo.Test
 
   import Plausible.Test.Support.HTML
+
+  @subject_prefix if ee?(), do: "[Plausible Analytics] ", else: "[Plausible CE] "
 
   setup [:create_user, :log_in]
 
@@ -115,7 +118,7 @@ defmodule PlausibleWeb.Site.MembershipControllerTest do
 
       assert_email_delivered_with(
         to: [nil: "john.doe@example.com"],
-        subject: "[Plausible Analytics] You've been invited to #{site.domain}"
+        subject: @subject_prefix <> "You've been invited to #{site.domain}"
       )
     end
 
@@ -130,7 +133,7 @@ defmodule PlausibleWeb.Site.MembershipControllerTest do
 
       assert_email_delivered_with(
         to: [nil: existing_user.email],
-        subject: "[Plausible Analytics] You've been invited to #{site.domain}"
+        subject: @subject_prefix <> "You've been invited to #{site.domain}"
       )
     end
 
@@ -168,7 +171,7 @@ defmodule PlausibleWeb.Site.MembershipControllerTest do
 
       assert_email_delivered_with(
         to: [nil: "joe@example.com"],
-        subject: "[Plausible Analytics] You've been invited to #{site.domain}"
+        subject: @subject_prefix <> "You've been invited to #{site.domain}"
       )
 
       req2 =
@@ -179,7 +182,7 @@ defmodule PlausibleWeb.Site.MembershipControllerTest do
 
       refute_email_delivered_with(
         to: [nil: "joe@example.com"],
-        subject: "[Plausible Analytics] You've been invited to #{site.domain}"
+        subject: @subject_prefix <> "You've been invited to #{site.domain}"
       )
 
       assert people_settings = redirected_to(req2, 302)
@@ -226,7 +229,7 @@ defmodule PlausibleWeb.Site.MembershipControllerTest do
 
       assert_email_delivered_with(
         to: [nil: "john.doe@example.com"],
-        subject: "[Plausible Analytics] Request to transfer ownership of #{site.domain}"
+        subject: @subject_prefix <> "Request to transfer ownership of #{site.domain}"
       )
     end
 
@@ -238,7 +241,7 @@ defmodule PlausibleWeb.Site.MembershipControllerTest do
 
       assert_email_delivered_with(
         to: [nil: existing_user.email],
-        subject: "[Plausible Analytics] Request to transfer ownership of #{site.domain}"
+        subject: @subject_prefix <> "Request to transfer ownership of #{site.domain}"
       )
     end
 
@@ -478,7 +481,7 @@ defmodule PlausibleWeb.Site.MembershipControllerTest do
 
       assert_email_delivered_with(
         to: [nil: admin.email],
-        subject: "[Plausible Analytics] Your access to #{site.domain} has been revoked"
+        subject: @subject_prefix <> "Your access to #{site.domain} has been revoked"
       )
     end
   end